+++ CHANGELOG.md
@@ -6,6 +6,10 @@ FEATURES:
 * builder/vmware: The root hard drive type can now be specified with
   "disk_type_id" for advanced users. [GH-328]
 
+IMPROVEMENTS:
+
+* builder/amazon: Tagging now works with all amazon builder types.
+
 BUG FIXES:
 
 * core: Fixed a couple cases where a double ctrl-C could panic.